adds registry to strict registration methods
