+++ src/background/partnerIntegrations.test.ts
@@ -33,7 +33,7 @@ import {
   secretsConfigFactory,
 } from "@/testUtils/factories/integrationFactories";
 import { appApiMock } from "@/testUtils/appApiMock";
-import { registry } from "@/background/messenger/api";
+import { registry } from "@/background/messenger/strict/api";
 import { setCachedAuthData } from "@/background/auth/authStorage";
 import {
   CONTROL_ROOM_OAUTH_INTEGRATION_ID,
