+++ src/bricks/registry.test.ts
@@ -16,7 +16,7 @@
  */
 
 import bricksRegistry from "@/bricks/registry";
-import { registry as backgroundRegistry } from "@/background/messenger/api";
+import { registry as backgroundRegistry } from "@/background/messenger/strict/api";
 import { echoBrick } from "@/runtime/pipelineTests/pipelineTestHelpers";
 import { parsePackage } from "@/registry/packageRegistry";
 import { starterBrickConfigFactory } from "@/testUtils/factories/modDefinitionFactories";
