+++ src/extensionConsole/pages/activateMod/IntegrationsBody.test.tsx
@@ -36,7 +36,8 @@ import {
   integrationDependencyFactory,
 } from "@/testUtils/factories/integrationFactories";
 import getModDefinitionIntegrationIds from "@/integrations/util/getModDefinitionIntegrationIds";
-import { registry, services } from "@/background/messenger/api";
+import { registry } from "@/background/messenger/strict/api";
+import { services } from "@/background/messenger/api";
 import { refreshServices } from "@/background/locator";
 import { clear, find, syncPackages } from "@/registry/packageRegistry";
 import { type ModDefinition } from "@/types/modDefinitionTypes";
